OPSMN-2156 : Issue with specimen quantity is resolved by changing datatype.
Description :
1. Specimen available quantity and initial quantity were double.
2. Double cannot handle decimal values appropriately.
3. Hence made changes in DB schema, so that bigDecimal can be stored for both the fields
   (decimal(16,6) for mysql and number(16,6) for oracle).
4. Same change is applied for initial quantity of specimen requirement and
   quantity of specimen order items.

+++ WEB-INF/src/com/krishagni/catissueplus/core/administrative/domain/DistributionOrderItem.java
@@ -1,5 +1,7 @@
 package com.krishagni.catissueplus.core.administrative.domain;
 
+import java.math.BigDecimal;
+
 import com.krishagni.catissueplus.core.biospecimen.domain.BaseEntity;
 import com.krishagni.catissueplus.core.biospecimen.domain.Specimen;
 
@@ -12,7 +14,7 @@ public class DistributionOrderItem extends BaseEntity {
 	
 	private DistributionOrder order;
 	
-	private Double quantity;
+	private BigDecimal quantity;
 	
 	private Specimen specimen;
 	
@@ -26,11 +28,11 @@ public class DistributionOrderItem extends BaseEntity {
 		this.order = order;
 	}
 
-	public Double getQuantity() {
+	public BigDecimal getQuantity() {
 		return quantity;
 	}
 
-	public void setQuantity(Double quantity) {
+	public void setQuantity(BigDecimal quantity) {
 		this.quantity = quantity;
 	}
